% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_queries.R
\name{divide_into_named_queries}
\alias{divide_into_named_queries}
\title{Divide Queries Into Separate Named Queries In List}
\usage{
divide_into_named_queries(content)
}
\arguments{
\item{content}{from clipboard.}
}
\value{
Named list.
}
\description{
Divide Queries Into Separate Named Queries In List
}
